First batch of n64 ABI commits for SGI (64-bit pointers, 64-bit longs, 32-bit ints).  More to come.
A few changes related to 64-bit pointer handling, but largely a compiler
warning cleaning pass for the SGI compilers (w/o -fullwarn).

Includes half a dozen bug fixes exposed by compiler warnings, though most
changes pertain to removing declared but not referenced, or initialized but
not used variables.

NOTES: More superfluous variables left in some of the
preprocessor-generated macros and .Z source files.  Also, left %.08x as the
format for printing pointers for now.

+++ dx/src/exec/libdx/edfdata.c
@@ -793,9 +793,8 @@ swap_64(void *src, void *dest, int n, int swaps)
     unsigned int *d = (unsigned int*)dest;
     unsigned int x,y,i;
     
-    if (((int)s & 3) || ((int)d & 3)) 
+    if (((long)s & 3) || ((long)d & 3)) 
 	DXErrorGoto(ERROR_INTERNAL,"Tried to swap non-word aligned data");
-    
     
     switch (swaps) {
       case (WORD_SWAP|HWORD_SWAP|BYTE_SWAP):
